+++ src/main/resources/template_create_tenant.sql
@@ -7,5 +7,6 @@ CREATE SCHEMA myuniversity_mymodule AUTHORIZATION myuniversity_mymodule;
 CREATE EXTENSION IF NOT EXISTS "pgcrypto";
 
 CREATE TABLE IF NOT EXISTS myuniversity_mymodule.auth_credentials (_id UUID PRIMARY KEY DEFAULT gen_random_uuid(), jsonb JSONB NOT NULL);
+CREATE TABLE IF NOT EXISTS myuniversity_mymodule.auth_attempts (_id UUID PRIMARY KEY DEFAULT gen_random_uuid(), jsonb JSONB NOT NULL);
 
 GRANT ALL PRIVILEGES ON ALL TABLES IN SCHEMA myuniversity_mymodule TO myuniversity_mymodule;